Move tests of units notn to `test_helpers.py`
